Velo linux upgrade fixes (#3122)
Encountered some issues testing the new Linux Upgrade artifacts and
created this PR as an attempt to address them. Issues observed:

1. Artifacts were not running due to precondition using capital 'L'.
Corrected precondition to lowercase.

2. The copy function coupled with Artifact.Generic.Utils.FetchBinary was
causing the resultant file to be empty, thus resulting in an error when
upgrade command was run. This is due to
Artifact.Generic.Utils.FetchBinary already storing the file in /tmp on
linux. Skipped by removing the copy.

3. Artifacts would error if tool (updated package) was already cached.
This is due to Artifact.Generic.Utils.FetchBinary returning an OSPath
object when the tool is cached vs returning a string when the tool is
downloaded. Fixed by explicitly casting `Dest` to str.

4. Running `rpm -U new-velo-package.rpm` directly on a shell on a host
would result in the package being upgraded but the velociraptor_client
service would be disabled and stopped. This is due to the upgrade
order-of-operations in RPM (see
https://docs.fedoraproject.org/en-US/packaging-guidelines/Scriptlets/#ordering).
Updates to rpm.go address this by only stopping/disabling service on
true uninstall vs an upgrade.

5. Debian upgrade artifact was loading the incorrect tool via
Artifact.Generic.Utils.FetchBinary. Changed to the correct tool.

6. Encountered a fundamental issue with the upgrade commands for both
Debian and RedHat in that the packages have pre-removal steps that stop
the velociraptor service which instantly terminates the upgrade command
( ie `rpm -U ` and `dpkg -i`) ; resulting in the upgrade command not
being successful. The commands were modified to skip any pre-removal
steps that stop the velociraptor service and then the service is
restarted to switch to the updated binary. This workaround, so to speak,
is the only way I have found to make velociraptor upgrade itself on
linux to work. On debian we delete the pre-removal script before
performing the upgrade and since with dpkg the old package is "removed"
before the new package is installed the pre-removal files get re-created
after the upgrade so the package can be properly uninstalled in the
future. On redhat with rpm we can just call a flag (`--nopreun`) to just
skip the pre-removal scriptlet; thus leaving it intact if the package
ever needs to be properly uninstalled.

Ref: (Deb order of operations
https://www.debian.org/doc/debian-policy/ap-flowcharts.html and RPM
order of operations
https://docs.fedoraproject.org/en-US/packaging-guidelines/Scriptlets/#ordering).

---------

+++ bin/rpm.go
@@ -317,9 +317,16 @@ func doClientRPM() error {
 /bin/systemctl start velociraptor_client.service
 `)
 
+	// check for upgrade vs uninstall
 	r.AddPreun(`
-/bin/systemctl disable velociraptor_client.service
-/bin/systemctl stop velociraptor_client.service
+if [ $1 == 1 ] ; then
+    /bin/systemctl restart velociraptor_client.service
+fi
+
+if [ $1 == 0 ] ; then
+    /bin/systemctl disable velociraptor_client.service
+    /bin/systemctl stop velociraptor_client.service
+fi
 `)
 
 	fd, err := os.OpenFile(output_path,
